Trim leading newline from app script
